refactor: moves http logic into a single module
