+++ ceramic-kubo-rpc/src/error.rs
@@ -1,11 +1,5 @@
 //! Provides simple error type for communicating specific
 //! HTTP status code along with errors.
-use actix_web::{
-    error,
-    http::{header::ContentType, StatusCode},
-    HttpResponse,
-};
-use serde::Serialize;
 use thiserror::Error;
 
 /// Error type for HTTP responses.
@@ -27,36 +21,4 @@ pub enum Error {
     // Represents the resource was not found.
     #[error("not found")]
     NotFound,
-}
-
-#[derive(Serialize)]
-struct ErrorJson<'a> {
-    #[serde(rename = "Message")]
-    pub message: String,
-    #[serde(rename = "Code")]
-    pub code: i32,
-    #[serde(rename = "Type")]
-    pub typ: &'a str,
-}
-
-impl error::ResponseError for Error {
-    fn error_response(&self) -> HttpResponse {
-        let err = ErrorJson {
-            message: self.to_string(),
-            code: 0,
-            typ: "error",
-        };
-        let data = serde_json::to_string(&err).unwrap();
-        HttpResponse::build(self.status_code())
-            .insert_header(ContentType::json())
-            .body(data)
-    }
-
-    fn status_code(&self) -> StatusCode {
-        match *self {
-            Error::InternalError(_) => StatusCode::INTERNAL_SERVER_ERROR,
-            Error::BadRequest(_) => StatusCode::BAD_REQUEST,
-            Error::NotFound => StatusCode::NOT_FOUND,
-        }
-    }
 }
