+++ ceramic-kubo-rpc/src/lib.rs
@@ -1,54 +1,21 @@
-use std::{collections::HashMap, net};
+use std::collections::HashMap;
 
-use actix_web::{web, App, HttpServer};
 use anyhow::Result;
 use async_trait::async_trait;
 use iroh_api::{Api, Bytes, Cid, IpfsPath, Multiaddr, PeerId};
-use tracing_actix_web::TracingLogger;
 
-mod dag;
-mod error;
-mod swarm;
-
-#[derive(Clone)]
-struct AppState<T>
-where
-    T: IrohClient,
-{
-    api: T,
-}
-
-/// Start the Kubo RPC mimic server.
-///
-/// Block until shutdown.
-/// Automatically registers shutdown listeners for interrupt and kill signals.
-/// See https://actix.rs/docs/server/#graceful-shutdown
-pub async fn serve<T, A>(api: T, addrs: A) -> std::io::Result<()>
-where
-    T: IrohClient + Send + Clone + 'static,
-    A: net::ToSocketAddrs,
-{
-    HttpServer::new(move || {
-        App::new()
-            .wrap(TracingLogger::default())
-            .app_data(web::Data::new(AppState { api: api.clone() }))
-            .service(
-                web::scope("/api/v0")
-                    .service(dag::scope::<T>())
-                    .service(swarm::scope::<T>()),
-            )
-    })
-    .bind(addrs)?
-    .run()
-    .await
-}
+pub mod dag;
+pub mod error;
+//TODO put behind feature
+pub mod http;
+pub mod swarm;
 
 /// Defines the behavior we need from Iroh in order to serve Kubo RPC calls.
 /// The trait serves two purposes:
 ///     1. We are explicit about the API surface area we consume from Iroh.
 ///     2. We can provide an implementation for testing.
 #[async_trait]
-pub trait IrohClient {
+pub trait IrohClient: Clone {
     type StoreClient: StoreClient;
     type P2pClient: P2pClient;
     fn try_store(&self) -> Result<Self::StoreClient>;
